requirements.txt
qdrant-client
sentence-transformers
rank-bm25
nltk
pandas
tqdm
matplotlib
pyyaml
numpy

